pyproject.toml
[tool.poetry]
name = "tournament-api"
version = "0.1.0"
description = "RESTful API for sport competitions: badminton, tennis, ping-pong, volleyball etc.\""
authors = ["Anastasiia Didan <anastasia.didan.dev@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.111.0"
fastapi-sqlalchemy = "^0.2.1"
alembic = "^1.13.1"
uvicorn = "^0.29.0"
asyncpg = "^0.29.0"
environs = "^11.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
pytest-mock = "^3.14.0"
pre-commit = "^3.3.3"
flake8 = "5.0"
black = "^24.4.2"
isort = "^5.13.2"
ipdb = "^0.13.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
